refactor: 날짜수 필드 형변환에 따른 수정

+++ src/main/java/com/backend/blooming/common/util/DayUtil.java
@@ -11,11 +11,11 @@ public class DayUtil {
 
     private static final int COUNT_GOAL_DAYS = 1;
 
-    public static int getDays(final LocalDate startDate, final LocalDate endDate) {
-        return (int) ChronoUnit.DAYS.between(startDate, endDate) + COUNT_GOAL_DAYS;
+    public static long getDays(final LocalDate startDate, final LocalDate endDate) {
+        return ChronoUnit.DAYS.between(startDate, endDate) + COUNT_GOAL_DAYS;
     }
 
-    public static int getNowDay(final LocalDate startDate) {
-        return (int) ChronoUnit.DAYS.between(startDate, LocalDate.now()) + COUNT_GOAL_DAYS;
+    public static long getNowDay(final LocalDate startDate) {
+        return ChronoUnit.DAYS.between(startDate, LocalDate.now()) + COUNT_GOAL_DAYS;
     }
 }
