+++ src/main/java/com/backend/blooming/stamp/presentation/dto/response/ReadStampResponse.java
@@ -6,7 +6,7 @@ public record ReadStampResponse(
         Long id,
         String userName,
         String userColor,
-        int day,
+        long day,
         String message
 ) {
